requirements1.txt
Flask-SQLAlchemy
psycopg2-binary
gunicorn
gevent
pytest-flask
plotly